demos/xrayedge/containers/image-server/requirements.txt
Flask==1.1.2
mysql-connector-python==8.0.*
flask-cors==3.0.8
